feat: allow osal to support multiple compilers
Instead of making this choice compile time,
install all supported compiler headers by default.
