+++ CMakeLists.txt
@@ -123,6 +123,17 @@ install(FILES
   DESTINATION include
   )
 
+install(FILES
+  include/sys/osal_cc_clang.h
+  include/sys/osal_cc_gcc.h
+  include/sys/osal_cc_msvc.h
+  include/sys/osal_cc_rtk.h
+  include/sys/osal_cc.h
+  include/osal_log.h
+  DESTINATION include
+  )
+
+
 if (CMAKE_PROJECT_NAME STREQUAL OSAL AND BUILD_TESTING)
   add_subdirectory (test)
   include(AddGoogleTest)
