+++ cmake/Windows.cmake
@@ -48,7 +48,6 @@ target_include_directories(osal PUBLIC
   )
 
 install(FILES
-  src/windows/sys/osal_cc.h
   src/windows/sys/osal_sys.h
   DESTINATION include/sys
   )
