+++ include/sys/osal_cc_gcc.h
@@ -22,19 +22,7 @@ extern "C" {
 
 #include <assert.h>
 
-#if defined(__clang__)
-#if !defined(CLANG_ANALYZER_NORETURN)
-#if __has_feature(attribute_analyzer_noreturn)
-#define CLANG_ANALYZER_NORETURN __attribute__ ((analyzer_noreturn))
-#else
-#define CLANG_ANALYZER_NORETURN
-#endif
-#endif
-#else
-#define CLANG_ANALYZER_NORETURN
-#endif
-
-static inline void cc_assert (int exp) CLANG_ANALYZER_NORETURN
+static inline void cc_assert (int exp)
 {
    assert (exp); // LCOV_EXCL_LINE
 }
